Port HSD allocators and pointer IDs to native memory

+++ src/sysdolphin/baselib/id.h
@@ -7,7 +7,7 @@
 
 typedef struct _IDEntry {
     struct _IDEntry* next;
-    u32 id;
+    uintptr_t id;
     void* data;
 } IDEntry;
 
@@ -20,16 +20,16 @@ void HSD_IDInitAllocData(void);
 void HSD_IDSetup(void);
 /// A NULL table selects the default table for insertion, removal, and lookup.
 /// Replaces the data for an existing ID without freeing the previous data.
-void HSD_IDInsertToTable(HSD_IDTable* table, u32 id, void* data);
+void HSD_IDInsertToTable(HSD_IDTable* table, uintptr_t id, void* data);
 /// Removes and frees the entry. The caller still owns its data.
-void HSD_IDRemoveByIDFromTable(HSD_IDTable* table, u32 id);
+void HSD_IDRemoveByIDFromTable(HSD_IDTable* table, uintptr_t id);
 /// Sets *success to 1 for a found ID, even when its data is NULL, or 0 if
 /// absent. The success output is optional.
-void* HSD_IDGetDataFromTable(HSD_IDTable* table, u32 id, s32* success);
+void* HSD_IDGetDataFromTable(HSD_IDTable* table, uintptr_t id, s32* success);
 /// Clears the entire default table without freeing entries. Ignores the range.
 void _HSD_IDForgetMemory(void* low, void* high);
 
-static inline void* HSD_IDGetData(u32 id, s32* success)
+static inline void* HSD_IDGetData(uintptr_t id, s32* success)
 {
     return HSD_IDGetDataFromTable(NULL, id, success);
 }
